+++ src/main/java/fuzs/pickupnotifier/config/ClientConfig.java
@@ -77,7 +77,7 @@ public class ClientConfig extends AbstractConfig {
 
         @Override
         protected void addToBuilder(ForgeConfigSpec.Builder builder, ConfigHolder.ConfigCallback saveCallback) {
-            saveCallback.accept(builder.comment("Disable specific items or content from whole mods from showing.", EntryCollectionBuilder.CONFIG_STRING).define("Blacklist", Lists.<String>newArrayList()), v -> this.blacklist = EntryCollectionBuilder.of(ForgeRegistries.ITEMS).buildSet(v));
+            saveCallback.accept(builder.comment("Disable specific items or content from whole mods from showing.", EntryCollectionBuilder.CONFIG_DESCRIPTION).define("Blacklist", Lists.<String>newArrayList()), v -> this.blacklist = EntryCollectionBuilder.of(ForgeRegistries.ITEMS).buildSet(v));
             saveCallback.accept(builder.comment("Combine entries of the same type instead of showing each one individually.").define("Combine Entries", true), v -> this.combineEntries = v);
             saveCallback.accept(builder.comment("Amount of ticks each entry will be shown for. Set to 0 to only remove entries when space for new ones is needed.").defineInRange("Display Time", 80, 0, Integer.MAX_VALUE), v -> this.displayTime = v);
             saveCallback.accept(builder.comment("Make outdated entries slowly move out of the screen instead of disappearing in place.").define("Move Out Of Screen", true), v -> this.move = v);